Move discretization and transforms modules
